authentication_service/requirements.txt
flask
flask-cors
psycopg2-binary
python-dotenv
werkzeug
bcrypt
